User commit

+++ src/main/java/com/example/demo/repository/ValidationRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import com.example.demo.entity.ValidationEntity;
 
 @Repository
-public interface ValidationRepository extends JpaRepository<ValidationEntity,Integer>{
+public interface ValidationRepository extends JpaRepository<ValidationEntity,Long>{
 
 }
